buildinfo related refactor: using the sama packages, structs as in the other projects
