+++ internal/app/productinfo/api/types.go
@@ -15,8 +15,6 @@
 package api
 
 import (
-	"runtime"
-
 	"github.com/banzaicloud/productinfo/pkg/productinfo"
 )
 
@@ -156,26 +154,5 @@ func NewServicesResponse(sds []productinfo.ServiceDescriber) ServicesResponse {
 	}
 	return ServicesResponse{
 		Services: services,
-	}
-}
-
-type BuildInfo struct {
-	Version    string `json:"version"`
-	CommitHash string `json:"commit_hash"`
-	BuildDate  string `json:"build_date"`
-	GoVersion  string `json:"go_version"`
-	Os         string `json:"os"`
-	Arch       string `json:"arch"`
-}
-
-// NewBuildInfo creates a new build info reference
-func NewBuildInfo(version, hash, date string) *BuildInfo {
-	return &BuildInfo{
-		Version:    version,
-		CommitHash: hash,
-		BuildDate:  date,
-		GoVersion:  runtime.Version(),
-		Os:         runtime.GOOS,
-		Arch:       runtime.GOARCH,
 	}
 }
